Fixed bugs and uploaded the necessary files.

+++ ALGI/trab1/pilha.h
@@ -1,7 +1,7 @@
 typedef struct pilha_dinamica PILHA_DINAMICA;
 
 typedef struct NO{
-    int *item;
+    long *item;
     struct NO *anterior;
 }NO;
 
@@ -14,8 +14,8 @@ struct pilha_dinamica{
 
 PILHA_DINAMICA *create_stack();
 NO *unit();
-void push(PILHA_DINAMICA *pilha, int x);
-int pop(PILHA_DINAMICA *pilha);
+void push(PILHA_DINAMICA *pilha, long x);
+long pop(PILHA_DINAMICA *pilha);
 void new_item(PILHA_DINAMICA *pilha, int x);
 void operation(PILHA_DINAMICA *pilha, char *op);
 void empty_stack(PILHA_DINAMICA *pilha);
